Move Ranger integration tests into Ranger extension
This change moves the ranger int-tests into the Ranger main module, because the reason for the separate project, the Quarkus build, no longer exists.
