+++ extensions/auth/ranger/impl/src/intTest/java/org/apache/polaris/extension/auth/ranger/test/RangerIntegrationTest.java
