+++ extensions/auth/ranger/impl/src/intTest/resources/authz_it_tests/dev_polaris_roles.json
