+++ gradle/projects.main.properties
@@ -50,7 +50,6 @@ polaris-extensions-federation-bigquery=extensions/federation/bigquery
 polaris-extensions-auth-opa=extensions/auth/opa/impl
 polaris-extensions-auth-opa-tests=extensions/auth/opa/tests
 polaris-extensions-auth-ranger=extensions/auth/ranger/impl
-polaris-extensions-auth-ranger-tests=extensions/auth/ranger/tests
 
 polaris-config-docs-annotations=tools/config-docs/annotations
 polaris-config-docs-generator=tools/config-docs/generator
